requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
agno==1.2.5
    # via -r requirements.in
altair==5.5.0
    # via streamlit
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   langfuse
    #   openai
    #   starlette
async-timeout==5.0.1
    # via redis
attrs==25.3.0
    # via
    #   jsonschema
    #   referencing
backoff==2.2.1
    # via langfuse
blinker==1.9.0
    # via streamlit
cachetools==5.5.2
    # via streamlit
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via
    #   duckduckgo-search
    #   streamlit
    #   typer
    #   uvicorn
deprecated==1.2.18
    # via redis
distro==1.9.0
    # via openai
docstring-parser==0.16
    # via agno
duckduckgo-search==7.5.5
    # via -r requirements.in
fastapi==0.115.12
    # via -r requirements.in
firecrawl==1.15.0
    # via -r requirements.in
gitdb==4.0.12
    # via gitpython
gitpython==3.1.44
    # via
    #   agno
    #   streamlit
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.7
    # via httpx
httpx==0.28.1
    # via
    #   agno
    #   langfuse
    #   openai
idna==3.10
    # via
    #   anyio
    #   httpx
    #   langfuse
    #   requests
jinja2==3.1.6
    # via
    #   altair
    #   pydeck
jiter==0.9.0
    # via openai
jsonschema==4.23.0
    # via altair
jsonschema-specifications==2024.10.1
    # via jsonschema
langfuse==2.60.2
    # via -r requirements.in
lxml==5.3.1
    # via duckduckgo-search
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
narwhals==1.32.0
    # via altair
nest-asyncio==1.6.0
    # via firecrawl
numpy==2.2.4
    # via
    #   pandas
    #   pydeck
    #   streamlit
openai==1.69.0
    # via -r requirements.in
packaging==24.2
    # via
    #   altair
    #   langfuse
    #   redis
    #   streamlit
pandas==2.2.3
    # via streamlit
pillow==11.1.0
    # via streamlit
primp==0.14.0
    # via duckduckgo-search
protobuf==5.29.4
    # via streamlit
pyarrow==19.0.1
    # via streamlit
pydantic==2.11.0
    # via
    #   -r requirements.in
    #   agno
    #   fastapi
    #   firecrawl
    #   langfuse
    #   openai
    #   pydantic-settings
pydantic-core==2.33.0
    # via pydantic
pydantic-settings==2.8.1
    # via agno
pydeck==0.9.1
    # via streamlit
pygments==2.19.1
    # via rich
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.1.0
    # via
    #   -r requirements.in
    #   agno
    #   firecrawl
    #   pydantic-settings
python-multipart==0.0.20
    # via agno
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via agno
redis==4.2.0
    # via -r requirements.in
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via
    #   firecrawl
    #   langfuse
    #   streamlit
rich==13.9.4
    # via
    #   agno
    #   typer
rpds-py==0.24.0
    # via
    #   jsonschema
    #   referencing
shellingham==1.5.4
    # via typer
six==1.17.0
    # via python-dateutil
smmap==5.0.2
    # via gitdb
sniffio==1.3.1
    # via
    #   anyio
    #   openai
starlette==0.46.1
    # via fastapi
streamlit==1.44.0
    # via -r requirements.in
tenacity==9.0.0
    # via streamlit
toml==0.10.2
    # via streamlit
tomli==2.2.1
    # via agno
tornado==6.4.2
    # via streamlit
tqdm==4.67.1
    # via openai
typer==0.15.2
    # via agno
typing-extensions==4.13.0
    # via
    #   agno
    #   altair
    #   anyio
    #   fastapi
    #   openai
    #   pydantic
    #   pydantic-core
    #   redis
    #   referencing
    #   streamlit
    #   typer
    #   typing-inspection
typing-inspection==0.4.0
    # via pydantic
tzdata==2025.2
    # via pandas
urllib3==2.3.0
    # via requests
uvicorn==0.34.0
    # via -r requirements.in
websockets==15.0.1
    # via firecrawl
wrapt==1.17.2
    # via
    #   deprecated
    #   langfuse
